Extracted custom Hook usePosts
